backend/Module9/TagTriviaQuestions/requirements.txt
google-cloud-firestore
google-cloud-language